requirements.txt
streamlit
mistralai
pdfplumber
python-docx
torch
transformers
faiss-cpu
python-dotenv
ollama
